Replace SonosUPnP with SonosCommandBuilder
Attempt to port the Java Sonos Controller of Valentin to C++ so that we can easily port functions from one library to another. GitHub repo of Sonos-Controller Android Library: https://github.com/vmichalak/sonos-controller

Technical note: Using std::string and std::map functions is not the favorized way to work on an embedded system, but it improves the code reusability and we have more than 40 kByte of heap available on the ESP8266. Speed is not actually a problem on a 80MHz microcontroller for such sporadic operations.