feature/sonos (#10)
* Replace SonosUPnP with SonosCommandBuilder
* Parse response with SonosResponseParser
* Update .travis.yml to explicitly including the ESP8266 2.4.0-rc2
* Use HttpClient
* Added new helper class XmlStringEscapeUtils to escape XML special characters
* Attempt to port the Java Sonos Controller of Valentin to C++ so that we can easily port functions from one library to another. GitHub repo of Sonos-Controller Android Library: https://github.com/vmichalak/sonos-controller

Technical note: Using std::string and std::map functions is not the favorized way to work on an embedded system, but it improves the code reusability and we have more than 40 kByte of heap available on the ESP8266. Speed is not actually a problem on a 80MHz microcontroller for such sporadic operations.